Cleaned up Flagged.tsx (Review is somewhat responsive), added helper function to utils.ts, add a new testing file for Flagged.tsx
