+++ client/src/utils/utils.ts
@@ -37,3 +37,20 @@ export function parseSearchResponseData(response: response_data) {
 
     return newClients;
 }
+
+// Helper function to find duplicates in the clients data
+export function findDuplicates(
+    clients: Array<{ [key: string]: string }>
+  ): Array<{ [key: string]: string }> {
+    const duplicates: Array<{ [key: string]: string }> = [];
+    const seen = new Set<string>();
+    clients.forEach((client) => {
+      const identifier = `${client["First Name"]} ${client["Last Name"]} ${client["Address"]} ${client["City"]} ${client["State"]}`;
+      if (seen.has(identifier)) {
+        duplicates.push({ ...client, Issue: "Duplicate Entry" });
+      } else {
+        seen.add(identifier);
+      }
+    });
+    return duplicates;
+}
